continous提速(cumsum问题); fp32问题

+++ chunk.py
@@ -28,7 +28,7 @@ class vanillaRWKV6(torch.autograd.Function):
             assert state.is_contiguous()
             eew = torch.exp(-torch.exp(w.float())).contiguous()
 
-            y = torch.empty((B, T, C), device=w.device, dtype=r.dtype, memory_format=torch.contiguous_format)
+            y = torch.empty((B, T, C), device=w.device, dtype=torch.float32, memory_format=torch.contiguous_format)
             if r.dtype == torch.bfloat16:
                 rwkv6.forward_bf16(B, T, C, H, state, r, k, v, eew, u, y)
             elif r.dtype == torch.float16:
@@ -73,7 +73,7 @@ class chunkRWKV6(torch.autograd.Function):
             assert state.is_contiguous()
 
             # 块内计算
-            y = torch.empty((B, nc, cs, H, HEAD_SIZE), device=w.device, dtype=r.dtype, memory_format=torch.contiguous_format) # result
+            y = torch.empty((B, nc, cs, H, HEAD_SIZE), device=w.device, dtype=torch.float32, memory_format=torch.contiguous_format) # result
             if r.dtype == torch.bfloat16:
                 rwkv6.forward_bf16(B*nc, cs, C, H, state, r, k, v, w, u, y)
             elif r.dtype == torch.float16:
